refactor cmd profile (#2073)
* refactor cmd profile

* more refactor
